gwt errors fixed

+++ artemis-gwt/src/main/java/com/artemis/backends/gwt/emu/com/artemis/Archetype.java
@@ -8,7 +8,7 @@ package com.artemis;
  */
 public final class Archetype {
 	final ComponentType[] types;
-	final int compositionId;
+	final short compositionId;
 
 	/**
 	 * @param types Desired composition of derived components.
@@ -16,6 +16,6 @@ public final class Archetype {
 	 */
 	public Archetype(ComponentType[] types, int compositionId) {
 		this.types = types;
-		this.compositionId = compositionId;
+		this.compositionId = (short) compositionId;
 	}
 }
